Rename depends-on-jts ot legacy-geo.
