+++ build.gradle
@@ -242,7 +242,7 @@ subprojects {
     "org.elasticsearch.plugin:parent-join-client:${version}": ':modules:parent-join',
     "org.elasticsearch.plugin:aggs-matrix-stats-client:${version}": ':modules:aggs-matrix-stats',
     "org.elasticsearch.plugin:percolator-client:${version}": ':modules:percolator',
-    "org.elasticsearch.plugin:depends-on-jts-client:${version}": ':modules:depends-on-jts',
+    "org.elasticsearch.plugin:legacy-geo-client:${version}": ':modules:legacy-geo',
   ]
   if (indexCompatVersions[-1].snapshot) {
     /* The last and second to last versions can be snapshots. Rather than use
