+++ client/transport/src/main/java/org/elasticsearch/transport/client/PreBuiltTransportClient.java
@@ -25,7 +25,7 @@ import org.elasticsearch.client.transport.TransportClient;
 import org.elasticsearch.common.SuppressForbidden;
 import org.elasticsearch.common.network.NetworkModule;
 import org.elasticsearch.common.settings.Settings;
-import org.elasticsearch.dependsonjts.DependsOnJtsPlugin;
+import org.elasticsearch.legacygeo.LegacyGeoPlugin;
 import org.elasticsearch.index.reindex.ReindexPlugin;
 import org.elasticsearch.join.ParentJoinPlugin;
 import org.elasticsearch.percolator.PercolatorPlugin;
@@ -90,7 +90,7 @@ public class PreBuiltTransportClient extends TransportClient {
                 PercolatorPlugin.class,
                 MustachePlugin.class,
                 ParentJoinPlugin.class,
-                DependsOnJtsPlugin.class));
+                LegacyGeoPlugin.class));
 
     /**
      * Creates a new transport client with pre-installed plugins.
