+++ modules/legacy-geo/licenses/jts-NOTICE.txt
