+++ modules/legacy-geo/licenses/lucene-NOTICE.txt
