+++ modules/legacy-geo/licenses/spatial4j-0.6.jar.sha1
