+++ modules/legacy-geo/licenses/spatial4j-NOTICE.txt
