+++ modules/legacy-geo/src/main/java/org/elasticsearch/common/geo/XShapeCollection.java
