+++ modules/legacy-geo/src/main/java/org/elasticsearch/common/geo/builders/CoordinateCollection.java
