+++ modules/legacy-geo/src/main/java/org/elasticsearch/common/geo/builders/LineStringBuilder.java
