+++ modules/legacy-geo/src/test/java/org/elasticsearch/common/geo/builders/GeometryCollectionBuilderTests.java
