+++ modules/legacy-geo/src/test/resources/org/elasticsearch/search/geo/gzippedmap.gz
