+++ qa/smoke-test-http/build.gradle
@@ -23,7 +23,7 @@ apply plugin: 'elasticsearch.test-with-dependencies'
 
 dependencies {
     testCompile project(path: ':modules:transport-netty4', configuration: 'runtime') // for http
-    testCompile project(path: ':modules:depends-on-jts', configuration: 'runtime')
+    testCompile project(path: ':modules:legacy-geo', configuration: 'runtime')
 }
 
 integTestRunner {
