+++ settings.gradle
@@ -31,11 +31,11 @@ List projects = [
   'test:logger-usage',
   'modules:aggs-matrix-stats',
   'modules:analysis-common',
-  'modules:depends-on-jts',
   'modules:ingest-common',
   'modules:lang-expression',
   'modules:lang-mustache',
   'modules:lang-painless',
+  'modules:legacy-geo',
   'modules:mapper-extras',
   'modules:parent-join',
   'modules:percolator',
